requirements.txt
numpy
Pillow
crc8
lzma
